Move enhance_construction_age_band due to circular import
