+++ asf_core_data/pipeline/preprocessing/feature_engineering.py
@@ -441,26 +441,6 @@ def get_building_entry_feature(df, feature):
     return df
 
 
-def enhance_construction_age_band(
-    construction_age_band: str,
-    transaction_type: str,
-) -> str:
-    """
-    Enhances construction age band by replacing unknown when transaction type is new dwelling
-    (EPC inspections only began in 2008, so if it's a new dwelling then the age band will be "2007 onwards").
-
-    Args:
-        construction_age_band: property's construction age band
-        transaction_type: transaction type (e.g. new dwelling)
-    Retruns:
-        Updated construction age band
-    """
-    if construction_age_band == "unknown":
-        if transaction_type == "new dwelling":
-            return "2007 onwards"
-    return construction_age_band
-
-
 def get_additional_features(df):
     """Add new features to the EPC dataset.
     The new features include information about the inspection and entry date,
